Passes the 4th test
Adds a new property to Task (ID), it sets the ID property in the constructor. Fixes some issues, gives ID when it tries to instantiate the Task.

The getById TaskList method returns null if the list is not set or the given ID is not in the list. Otherwise it returns the Task object with the specific ID.