Passes all the remaining test
Because the task didn't ask it specifically to commit after each sub-task, I just didn't do it so here is everything I made.

SortBy{Name,ID} methods:
It uses the usort method which is kind of similar to the built in sort method, but instead of rearrange the array's keys, it assigns new keys to the value in the array and it uses a callback function, so we can use to to sort the array based on any value basically, we just have to write the business logic to do it and we have to make sure that we always return a new index if we are working with non-associative arrays.